refactor: TaskFormFields 추출 + TaskCreateModal 재구성
기능 C에서 같은 폼을 EditModal과 공유하기 위한 사전 리팩터링. 동작 변경 없음.

- components/tasks/task-form-fields.tsx (신규):
  - 기존 task-create-modal에 있던 useReducer + 7필드 + 검증(titleError·dateError·progressError) 그대로 이전
  - Dialog.Body + Dialog.Footer fragment 반환 → 부모 Dialog.Content 안에서 사용
  - props: initialValues / submitting / onSubmit
- components/tasks/task-create-modal.tsx:
  - Dialog 외곽만 남기고 폼은 TaskFormFields에 위임
  - initialValues={EMPTY_FORM_STATE}로 빈 값 시작
  - Dialog.Content key={open?'open':'closed'} 트릭으로 닫힐 때 자연스럽게 reset

기능 B의 J2/J3/J17/필수입력/진행률 회귀가 깨지지 않게 유지.
